requirements.txt
streamlit==1.32.0
flask==3.0.2
flask-cors==4.0.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-google-genai==0.0.11
langchain-groq==0.1.4
langchain-community==0.0.27
python-dotenv==1.0.1
Pillow==10.2.0
google-generativeai==0.3.2
groq==0.4.2
pydantic==2.6.3
requests==2.31.0
python-multipart==0.0.9
numpy==1.26.4
pandas==2.2.1
python-jose==3.3.0
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
mypy==1.8.0
httpx==0.27.0
python-json-logger==2.0.7
tenacity==8.2.3 